Chore: Bump `svelte` for tests (#121)
* chore: bump `svelte` version

* chore: update CSS `sources` expectants

* Chore: Attach `uvu` runner (#122)
